requirements.txt
google-genai==1.73.0
gradio==6.12.0
python-dotenv==1.2.2
tenacity==9.1.4
httpx==0.28.1
pydantic==2.13.0
fastapi==0.135.3
uvicorn==0.44.0
starlette==1.0.0
numpy==2.4.4
pandas==3.0.2
pillow==12.2.0
requests==2.33.1
